+++ FastSocket/FastSocket.swift
@@ -15,28 +15,25 @@ import Network
 /// low level TCP communication protocol to measure TCP throughput performance. -> FastSocket is the answer
 /// FastSocket allows to enter all possible TCP Options if needed and is completely non-blocking and async, thanks to GCD
 public final class FastSocket: FastSocketProtocol {
-    public var on = SocketCallback()
+    public var on = FastSocketCallback()
     public var parameters = TransferParameters()
     private var host: String
     private var port: UInt16
     private var frame = Frame()
     private var transfer: TransferProtocol?
     private var timer: DispatchSourceTimer?
-    private var sha256 = Data()
+    private var digest = Data()
     private var type: TransferType
     private var isLocked = false
-    private var allowUntrusted = false
     /// create a instance of FastSocket
     /// - parameters:
     ///     - host: a server endpoint to connect, e.g.: "example.com"
     ///     - port: the port to connect, e.g.: 8000
     ///     - type: the transfer type (.tcp or .tls)
-    ///     - allowUntrusted: if .tls connection are set, then allow untrusted certs
-    public required init(host: String, port: UInt16, type: TransferType = .tcp, allowUntrusted: Bool = false) {
+    public required init(host: String, port: UInt16, type: TransferType = .tcp) {
         self.host = host
         self.port = port
         self.type = type
-        self.allowUntrusted = allowUntrusted
     }
     /// connect to the server
     /// try to establish a connection to a
@@ -61,7 +58,7 @@ public final class FastSocket: FastSocketProtocol {
     /// generic send function, send data or string based messages
     /// - parameters:
     ///     - message: generic type (accepts data or string)
-    public func send<T: MessageTypeProtocol>(message: T) {
+    public func send<T: MessageProtocol>(message: T) {
         guard isLocked, let transfer = transfer else {
             return
         }
@@ -83,10 +80,9 @@ private extension FastSocket {
             return
         }
         isLocked = false
-        sha256 = Data()
-        transfer = NetworkTransfer(host: host, port: port, type: type, allowUntrusted: allowUntrusted, parameters: parameters)
+        digest = Data()
+        transfer = NetworkTransfer(host: host, port: port, type: type, parameters: parameters)
         transferCallbacks()
-        frameCallbacks()
     }
     /// suspends timeout and report on error
     /// - parameters:
@@ -110,7 +106,7 @@ private extension FastSocket {
             onError(FastSocketError.handshakeInitializationFailed)
             return
         }
-        sha256 = data.sha256
+        digest = data.sha256
         transfer.send(data: data)
     }
     /// start timeout on connecting
@@ -132,7 +128,7 @@ private extension FastSocket {
     /// closures from the transfer protocol
     /// handles incoming data and handshake
     private func transferCallbacks() {
-        guard var transfer = transfer else {
+        guard let transfer = transfer else {
             return
         }
         transfer.on.ready = { [weak self] in
@@ -151,11 +147,6 @@ private extension FastSocket {
         transfer.on.close = on.close
         transfer.on.bytes = on.bytes
     }
-    /// closures from Frame
-    /// returns the parsed messages
-    private func frameCallbacks() {
-        frame.onMessage = on.message
-    }
 }
 
 private extension FastSocket {
@@ -168,20 +159,21 @@ private extension FastSocket {
     /// this function is called from
     /// the transfer, to handle all necessary
     /// things `on message`
-    private func handleMessageState(data: MessageTypeProtocol) {
+    private func handleMessageState(data: MessageProtocol) {
         guard case let data as Data = data else {
             return
         }
         switch self.isLocked {
         case true:
             do {
-                try self.frame.parse(data: data)
+                try self.frame.parse(data: data) { message in
+                    on.message(message)
+                }
             } catch {
                 self.onError(error)
             }
-
         case false:
-            guard data == self.sha256 else {
+            guard data == self.digest else {
                 self.onError(FastSocketError.handshakeVerificationFailed)
                 return
             }