+++ FastSocket/Model/ByteCountResult.swift
@@ -5,14 +5,13 @@
 //  Created by Vinzenz Weist on 15.05.19.
 //  Copyright © 2019 Vinzenz Weist. All rights reserved.
 //
-
 import Foundation
 /// generic result type for transmitted bytes
-public enum ByteCountResult<Bytes> {
+public enum ByteCountResult {
     /// input bytes are the bytes
     /// which are *readed from the socket*
-    case input(Bytes)
+    case input(Int)
     /// output bytes are the bytes
     /// which are *written on the socket*
-    case output(Bytes)
+    case output(Int)
 }