+++ FastSocket/Transfer/NetworkTransfer.swift
@@ -12,8 +12,8 @@ import Network
 /// it uses the Network.framework. This is
 /// the `Engine` of the FastSocket Protocol.
 /// It allows to enter directly the TCP Options
-internal class NetworkTransfer: TransferProtocol {
-    internal var on = SocketCallback()
+internal final class NetworkTransfer: TransferProtocol {
+    internal var on = FastSocketCallback()
     private var connection: NWConnection?
     private var monitor = NWPathMonitor()
     private var parameters: TransferParameters
@@ -23,32 +23,6 @@ internal class NetworkTransfer: TransferProtocol {
     private var queue: DispatchQueue
     private var type: TransferType
     private var isRunning: Bool = false
-    private var isLocked: Bool = false
-    private var allowUntrusted: Bool = false
-    /// computed property to overwrite standard
-    /// tls options to give the ability to allow untrusted certs
-    /// - returns: NWProtocolTLS.Options object
-    var options: NWProtocolTLS.Options {
-        let options = NWProtocolTLS.Options()
-        sec_protocol_options_set_verify_block(options.securityProtocolOptions, { _, secTrust, secProtocolVerifyComplete in
-            let trust = sec_trust_copy_ref(secTrust).takeRetainedValue()
-            var error: CFError?
-            switch SecTrustEvaluateWithError(trust, &error) {
-            case true:
-                secProtocolVerifyComplete(true)
-
-            case false:
-                switch self.allowUntrusted {
-                case true:
-                    secProtocolVerifyComplete(true)
-
-                case false:
-                    secProtocolVerifyComplete(false)
-                }
-            }
-        }, queue)
-        return options
-    }
     /// computed property to create the right transfer type
     /// and mapps the allowed parameters to the NWParameters
     /// - returns: NWParameters object
@@ -59,7 +33,7 @@ internal class NetworkTransfer: TransferProtocol {
             param = NWParameters(tls: nil)
 
         case .tls:
-            param = NWParameters(tls: options)
+            param = NWParameters(tls: .init())
         }
         param.acceptLocalOnly = parameters.acceptLocalOnly
         param.allowFastOpen = parameters.allowFastOpen
@@ -76,14 +50,12 @@ internal class NetworkTransfer: TransferProtocol {
     ///     - host: a server endpoint to connect, e.g.: "example.com"
     ///     - port: the port to connect, e.g.: 8000
     ///     - type: the transfer type (.tcp or .tls)
-    ///     - allowUntrusted: if .tls connection are set, then allow untrusted certs
     ///     - transferParameters: TransferParameters `optional`
     ///     - queue: Dispatch Qeue `optional`
-    required init(host: String, port: UInt16, type: TransferType = .tcp, allowUntrusted: Bool = false, parameters: TransferParameters = TransferParameters(), queue: DispatchQueue = DispatchQueue(label: "\(Constant.prefixNetwork)\(UUID().uuidString)", qos: .userInitiated)) {
+    required init(host: String, port: UInt16, type: TransferType = .tcp, parameters: TransferParameters = TransferParameters(), queue: DispatchQueue = DispatchQueue(label: "\(Constant.prefixNetwork)\(UUID().uuidString)", qos: .userInitiated)) {
         self.host = host
         self.port = port
         self.type = type
-        self.allowUntrusted = allowUntrusted
         self.parameters = parameters
         self.queue = queue
     }
@@ -121,27 +93,17 @@ internal class NetworkTransfer: TransferProtocol {
             guard let self = self else {
                 return
             }
-            // this code prevent from multiple send of data at once
-            // this feature is now possible, but in release versions it will be blocked for now
-            // later we will allow this after loooooooooot of testing
-            #if !DEBUG
-            guard !isLocked else {
-                on.error(FastSocketError.writeBeforeClear)
-                return
-            }
-            #endif
             guard connectionState == .ready else {
                 on.error(FastSocketError.sendToEarly)
                 return
             }
-            isLocked = true
             let queued = data.chunk(by: Constant.iterations)
             guard !queued.isEmpty else {
                 return
             }
-            var iterator = queued.enumerated().makeIterator()
-            while let (index, data) = iterator.next() {
-                connection.send(content: Data(data), completion: .contentProcessed({ error in
+            var iterator = queued.makeIterator()
+            while let data = iterator.next() {
+                connection.send(content: data, completion: .contentProcessed({ error in
                     if let error = error {
                         guard error != NWError.posix(.ECANCELED) else {
                             // cancel error can be ignored
@@ -151,9 +113,6 @@ internal class NetworkTransfer: TransferProtocol {
                         return
                     }
                     self.on.bytes(.output(data.count))
-                    if index == queued.endIndex - 1 {
-                        self.isLocked = false
-                    }
                 }))
             }
         }